Migration to ATmega + touch-ups
Everything has been migrated to an atmega microcontroller instead of the arduino uno board. The only connections to the uno now are the transmit and receive (for programming the atmega) and the 5v power supply from the arduino since I don't yet have batteries that will provide 5v. 
Additionally, I did work on the game to iron out a lot of the issues. I have made the gameplay difficulty way better, added a secret command to reset the all time highscore, and a lot of minor improvements. There is still more to do though.